Add files via upload
This version is almost a replicate of the Pan-Can file Dr. Ferrall-Fairbanks created. The code was only changed on the loss function to reflect the extra time points I had. 
For the Tyk-nu cell line, this code works better than for Tyk-nu cp.r GFP. The problem with this code is that the carrying capacity estimated is too high, meaning it is almost the value that we selected as the top boundary for the carrying capacity. This is true for logistic, gompertz, and generalized logistic growth models. 
Using this code, the best growth models are von bertanlanffy and exponential for Tyk-nu and Tyk-nu cp.r GFP, respectively.