Tyk-nu and Tyk-nu cpr GFP co-culture code
This code was used to calculate for the sensitive to resistant conversion rates in Tyk-nu and Tyk-nu cpr GFP co-cultures. Here cell growth parameters and carrying capacities were  calculated from the optimization.ipynb file from the logistic growth model. This code work but the conversion rate values are not realistic. This was presented on the BMES poster.