Tyk-nu and Tyk-nu cpr GFP co-culture same PanCan
This code is similar to the PanCan code but the loss function and initial conditions were changed to fit the data for the two cell lines. Here the growth rate parameters and carrying capacities came from the code Pop dynamics Tyk-nu Sep2023.ipynb for the logistic growth model. In this code, I tried to calculate for only ONE conversion rate between sensitive to resistant. This code runs very slowly and sometimes it crashes before it finishes running the optimization problem. Once it is done the conversion rate is really small (almost 0) and the error is really high.