New AIC function created for all cell lines
I created this function to be able to iterate through the replicates. Here the loss function used is the residual sum of squares. This code runs and it does not need boundaries for the optimization problem to find the right parameters. The estimated curves fit the data pretty well but the AIC values are really high. This code was used for all the cell lines and the five growth models represented on the UFHCC 2024 poster.